+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 10.0.0
+## 1.0.0
 
 - Updates to .NET 10 and Aspire 13
+
+## 1.0.1
+
+- Fixes multiple providers in the same category mixmatching service registration